fix: props from TaskView reduce
